PRUEBAS HECHAS

+++ database/migrations/2025_10_31_174517_add_codigo_unico_secciones_table.php
@@ -23,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('seccions', function (Blueprint $table) {
-            $table->dropColumn('codigo');
-        });
+        if (Schema::hasTable('seccions')) {
+            Schema::table('seccions', function (Blueprint $table) {
+                $table->dropColumn('codigo');
+            });
+        }
     }
 };
